requirements.txt
asgiref==3.2.7
autopep8==1.5.7
Django==3.0.5
djangorestframework==3.12.4
pycodestyle==2.7.0
pytz==2019.3
sqlparse==0.3.1
toml==0.10.2
